Create detection.js
A hacky way to implement 3D mouse detection in this library. I am not happy with this, but I was not able to think of anything else at this point in time. It can be improved, especially in terms of how colors are assigned and detected, as right now it only supports 16x16 mine games. But that is a task for another time.